pyproject.toml
[project]
name = "agentlens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "ollama>=0.6.1",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]

